pyproject.toml
[tool.poetry]
name = "solar-backend"
version = "0.1.0"
description = ""
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/gabriel-milan/solar-backend"
repository = "https://github.com/gabriel-milan/solar-backend"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
tortoise-orm = { version = "^0.20.0", extras = ["asyncpg"] }
uvicorn = { version = "^0.24.0.post1", extras = ["standard"] }
loguru = "^0.7.2"
aerich = "^0.7.2"
fastapi-pagination = "^0.12.12"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
pre-commit = "^2.18.1"
taskipy = "^1.11.0"
isort = "^5.12.0"
semgrep = "^1.34.1"

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 100
target-version = ["py311"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
lint = "semgrep scan --error --verbose --config auto && black . && isort . && flake8 ."
serve = "uvicorn app.main:app --reload --port 8001"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
